requirements.txt
flask
gunicorn
python-dotenv
openai
gnews
feedparser
psycopg2-binary
